update + add test
