+++ frontend/packages/ux-editor/src/components/Properties/PropertiesHeader/ComponentMainConfig.test.tsx
@@ -3,10 +3,8 @@ import type { FormItem } from '@altinn/ux-editor/types/FormItem';
 import { ComponentType } from 'app-shared/types/ComponentType';
 import { screen } from '@testing-library/react';
 import { ComponentMainConfig } from './ComponentMainConfig';
-import { component1Mock } from '@altinn/ux-editor/testing/layoutMock';
-import { addFeatureFlagToLocalStorage, FeatureFlag } from 'app-shared/utils/featureToggleUtils';
 import { textMock } from '@studio/testing/mocks/i18nMock';
-import { typedLocalStorage } from '@studio/pure-functions';
+import { component1Mock } from '@altinn/ux-editor/testing/layoutMock';
 import { renderWithProviders } from '../../../testing/mocks';
 import { createQueryClientMock } from 'app-shared/mocks/queryClientMock';
 import { QueryKey } from 'app-shared/types/QueryKey';
@@ -21,25 +19,18 @@ const summary2ComponentMock: FormItem = {
 };
 
 describe('ComponentMainConfig', () => {
-  afterEach(() => {
-    jest.clearAllMocks();
-    typedLocalStorage.removeItem('featureFlags');
-  });
+  afterEach(() => jest.clearAllMocks);
 
   it('should render summary2 config when the component type matches', async () => {
     renderComponentMainConfig(summary2ComponentMock);
-
     const targetHeader = screen.getByText(textMock('ux_editor.component_properties.target'));
     expect(targetHeader).toBeInTheDocument();
   });
 
-  it('should render header config when feature flag is set, but the type does not match', async () => {
-    addFeatureFlagToLocalStorage(FeatureFlag.MainConfig);
+  it('should not render any config when the component type does not match', async () => {
     renderComponentMainConfig(component1Mock);
-
-    const sectionHeader = textMock('ux_editor.component_properties.main_configuration');
-    const headerMainConfig = screen.getByText(sectionHeader);
-    expect(headerMainConfig).toBeInTheDocument();
+    const wrapper = screen.getByTestId('component-wrapper');
+    expect(wrapper).toBeEmptyDOMElement();
   });
 });
 
@@ -48,7 +39,9 @@ const renderComponentMainConfig = (component: FormItem) => {
   const queryClient = createQueryClientMock();
   queryClient.setQueryData([QueryKey.LayoutSetsExtended, org, app], layoutSetsExtendedMock);
   return renderWithProviders(
-    <ComponentMainConfig component={component} handleComponentChange={handleComponentChange} />,
+    <div data-testid='component-wrapper'>
+      <ComponentMainConfig component={component} handleComponentChange={handleComponentChange} />
+    </div>,
     { queryClient },
   );
 };
