+++ frontend/packages/ux-editor/src/components/Properties/PropertiesHeader/PropertiesHeader.test.tsx
@@ -16,6 +16,8 @@ import type { IFormLayouts } from '@altinn/ux-editor/types/global';
 import { app, org } from '@studio/testing/testids';
 import { ComponentType } from 'app-shared/types/ComponentType';
 import { componentMocks } from '@altinn/ux-editor/testing/componentMocks';
+import { addFeatureFlagToLocalStorage, FeatureFlag } from 'app-shared/utils/featureToggleUtils';
+import { typedLocalStorage } from '@studio/pure-functions';
 
 const mockHandleComponentUpdate = jest.fn();
 
@@ -30,7 +32,10 @@ const defaultProps: PropertiesHeaderProps = {
 };
 
 describe('PropertiesHeader', () => {
-  afterEach(jest.clearAllMocks);
+  afterEach(() => {
+    jest.clearAllMocks();
+    typedLocalStorage.removeItem('featureFlags');
+  });
 
   it('renders the header name for the component', () => {
     renderPropertiesHeader();
@@ -162,7 +167,24 @@ describe('PropertiesHeader', () => {
     const alert = screen.queryByText(textMock('ux_editor.component_properties.deprecated.Input'));
     expect(alert).not.toBeInTheDocument();
   });
+
+  it('should render main configuration header when feature flag is set', () => {
+    addFeatureFlagToLocalStorage(FeatureFlag.MainConfig);
+    renderPropertiesHeader({ formItem: componentMocks[ComponentType.Input] });
+
+    const sectionHeader = textMock('ux_editor.component_properties.main_configuration');
+    const headerMainConfig = screen.getByText(sectionHeader);
+    expect(headerMainConfig).toBeInTheDocument();
+  });
+
+  it('should not render main configuration header when feature flag is not set', () => {
+    renderPropertiesHeader({ formItem: componentMocks[ComponentType.Input] });
+    const sectionHeader = textMock('ux_editor.component_properties.main_configuration');
+    const headerMainConfig = screen.queryByText(sectionHeader);
+    expect(headerMainConfig).not.toBeInTheDocument();
+  });
 });
+
 const renderPropertiesHeader = (props: Partial<PropertiesHeaderProps> = {}) => {
   const componentType = props.formItem ? props.formItem.type : defaultProps.formItem.type;
   queryClientMock.setQueryData(
